Moved GeoDataFrameWrapper interface to coordinate_clustering
