+++ src/sensai/evaluation/clustering_ground_truth.py
@@ -9,7 +9,7 @@ from geopandas import GeoDataFrame
 from shapely.geometry import Polygon, MultiPoint, MultiPolygon
 from typing import Sequence, Union, Optional
 
-from ..base.interfaces import GeoDataFrameWrapper
+from ..clustering.coordinate_clustering import GeoDataFrameWrapper
 from ..util.coordinates import validateCoordinates
 
 log = logging.getLogger(__name__)
